feat(backend): re factored question module
