+++ backend/src/modules/core/classes/transformers/Question.ts
@@ -1,79 +1,79 @@
-import {
-  ObjectIdToString,
-  StringToObjectId,
-} from '#shared/constants/transformerConstants.js';
-import {
-  IQuestion,
-  IQuestionMetrics,
-  IQuestionPriority,
-  QuestionStatus,
-} from '#shared/interfaces/models.js';
-import {Expose, Transform} from 'class-transformer';
-import {ObjectId} from 'mongodb';
+// import {
+//   ObjectIdToString,
+//   StringToObjectId,
+// } from '#shared/constants/transformerConstants.js';
+// import {
+//   IQuestion,
+//   IQuestionMetrics,
+//   IQuestionPriority,
+//   QuestionStatus,
+// } from '#shared/interfaces/models.js';
+// import {Expose, Transform} from 'class-transformer';
+// import {ObjectId} from 'mongodb';
 
-class Question implements IQuestion {
-  @Transform(ObjectIdToString.transformer, {toPlainOnly: true})
-  @Transform(StringToObjectId.transformer, {toClassOnly: true})
-  @Expose()
-  _id?: string | ObjectId;
+// class Question implements IQuestion {
+//   @Transform(ObjectIdToString.transformer, {toPlainOnly: true})
+//   @Transform(StringToObjectId.transformer, {toClassOnly: true})
+//   @Expose()
+//   _id?: string | ObjectId;
 
-  @Transform(ObjectIdToString.transformer, {toPlainOnly: true})
-  @Transform(StringToObjectId.transformer, {toClassOnly: true})
-  @Expose()
-  userId: string | ObjectId;
+//   @Transform(ObjectIdToString.transformer, {toPlainOnly: true})
+//   @Transform(StringToObjectId.transformer, {toClassOnly: true})
+//   @Expose()
+//   userId: string | ObjectId;
 
-  @Expose()
-  question: string;
+//   @Expose()
+//   question: string;
 
-  @Transform(ObjectIdToString.transformer, {toPlainOnly: true})
-  @Transform(StringToObjectId.transformer, {toClassOnly: true})
-  @Expose()
-  context: string | ObjectId;
+//   @Transform(ObjectIdToString.transformer, {toPlainOnly: true})
+//   @Transform(StringToObjectId.transformer, {toClassOnly: true})
+//   @Expose()
+//   context: string | ObjectId;
 
-  @Expose()
-  status: QuestionStatus;
+//   @Expose()
+//   status: QuestionStatus;
 
-  @Expose()
-  totalAnswersCount: number;
+//   @Expose()
+//   totalAnswersCount: number;
 
-  @Expose()
-  embedding: number[];
+//   @Expose()
+//   embedding: number[];
 
-  @Expose()
-  priority: IQuestionPriority;
+//   @Expose()
+//   priority: IQuestionPriority;
 
-  @Expose()
-  metrics: IQuestionMetrics | null;
+//   @Expose()
+//   metrics: IQuestionMetrics | null;
   
-  @Expose()
-  text: string;
+//   @Expose()
+//   text: string;
 
-  @Expose()
-  details: {
-    state: string;
-    district: string;
-    crop: string;
-    season: string;
-    domain: string;
-  };
+//   @Expose()
+//   details: {
+//     state: string;
+//     district: string;
+//     crop: string;
+//     season: string;
+//     domain: string;
+//   };
 
-  @Expose()
-  source: 'AJRASAKHA' | 'AGRI_EXPERT';
+//   @Expose()
+//   source: 'AJRASAKHA' | 'AGRI_EXPERT';
 
-  @Expose()
-  isAutoAllocate: boolean;
+//   @Expose()
+//   isAutoAllocate: boolean;
 
-  @Expose()
-  createdAt?: Date;
+//   @Expose()
+//   createdAt?: Date;
 
-  @Expose()
-  updatedAt?: Date;
+//   @Expose()
+//   updatedAt?: Date;
 
-  constructor(data?: Partial<IQuestion>) {
-    // Object.assign(this, data);
-    this.createdAt = data?.createdAt ?? new Date();
-    this.updatedAt = data?.updatedAt ?? new Date();
-  }
-}
+//   constructor(data?: Partial<IQuestion>) {
+//     // Object.assign(this, data);
+//     this.createdAt = data?.createdAt ?? new Date();
+//     this.updatedAt = data?.updatedAt ?? new Date();
+//   }
+// }
 
-export {Question};
+// export {Question};
