+++ backend/src/modules/core/index.ts
@@ -2,10 +2,8 @@ import {sharedContainerModule} from '#root/container.js';
 import {Container, ContainerModule} from 'inversify';
 import {InversifyAdapter} from '#root/inversify-adapter.js';
 import {useContainer} from 'class-validator';
-import {QuestionController} from './controllers/QuestionController.js';
 import {ContextController} from './controllers/ContextController.js';
 import {coreContainerModule} from './container.js';
-import {QUESTION_VALIDATORS} from './classes/validators/QuestionValidators.js';
 import {CONTEXT_VALIDATORS} from './classes/validators/ContextValidators.js';
 import {COMMENT_VALIDATORS} from './classes/validators/CommentValidators.js';
 import {CommentController} from './controllers/CommentController.js';
@@ -18,6 +16,8 @@ import { NOTIFICATION_VALIDATORS } from './classes/validators/NotificationValida
 import {PerformanceController} from './controllers/PerformanceController.js'
 import { AnswerController } from '../answer/controllers/AnswerController.js';
 import { ANSWER_VALIDATORS } from '../answer/classes/validators/AnswerValidator.js';
+import { QuestionController } from '../question/controllers/QuestionController.js';
+import { QUESTION_VALIDATORS } from '../question/classes/validators/QuestionVaidators.js';
 
 // Export names that loadAppModules expects
 export const coreModuleControllers: Function[] = [
@@ -56,14 +56,14 @@ export const coreModuleValidators: Function[] = [
 ];
 
 // Export all the main components for external use
-export * from './controllers/QuestionController.js';
+export * from '../../modules/question/controllers/QuestionController.js';
 export * from '../../modules/answer/controllers/AnswerController.js'
 export * from './controllers/ContextController.js';
 export * from './controllers/CommentController.js';
 export * from './controllers/NotificationController.js'
 export * from './controllers/PerformanceController.js'
 
-export * from './services/QuestionService.js';
+export * from '../../modules/question/services/QuestionService.js';
 export * from '../../modules/answer/services/AnswerService.js'
 export * from './services/ContextService.js';
 export * from './services/CommentService.js';
