+++ backend/src/modules/core/services/PerformanceService.ts
@@ -5,31 +5,13 @@ import {GLOBAL_TYPES} from '#root/types.js';
 import {inject, injectable} from 'inversify';
 import {ClientSession, ObjectId} from 'mongodb';
 import {
-  IAnswer,
-  IQuestionMetrics,
-  ISubmissionHistory,
   IReviewerHeatmapRow,
-  QuestionStatus,
 } from '#root/shared/interfaces/models.js';
 import {
-  BadRequestError,
-  InternalServerError,
-  NotFoundError,
   UnauthorizedError,
 } from 'routing-controllers';
-import {CORE_TYPES} from '../types.js';
-import {AiService} from './AiService.js';
 import {IQuestionSubmissionRepository} from '#root/shared/database/interfaces/IQuestionSubmissionRepository.js';
-import {dummyEmbeddings} from '../utils/questionGen.js';
-import {
-  IQuestionAnalysis,
-  IQuestionWithAnswerTexts,
-} from '../classes/validators/QuestionValidators.js';
-import {QuestionService} from './QuestionService.js';
 import {IUserRepository} from '#root/shared/database/interfaces/IUserRepository.js';
-import {INotificationRepository} from '#root/shared/database/interfaces/INotificationRepository.js';
-import {notifyUser} from '#root/utils/pushNotification.js';
-import {NotificationService} from './NotificationService.js';
 import {
   Analytics,
   AnswerStatusOverview,
