+++ backend/src/shared/database/interfaces/IQuestionRepository.ts
@@ -10,10 +10,7 @@ import {
   ModeratorApprovalRate,
   QuestionStatusOverview,
 } from '#root/modules/core/classes/validators/DashboardValidators.js';
-import {
-  GetDetailedQuestionsQuery,
-  QuestionResponse,
-} from '#root/modules/core/classes/validators/QuestionValidators.js';
+import { GetDetailedQuestionsQuery, QuestionResponse } from '#root/modules/question/classes/validators/QuestionVaidators.js';
 import {
   IQuestion,
   IUser,
